[REFACTOR]: move session endpoints and tests into api app.
