+++ backend/api/views.py
@@ -22,6 +22,17 @@ from artifact_history.services import (
     update_artifact_history_custom_name,
 )
 from authentication.permissions import IsVerifiedUser
+from chat_sessions.serializers import (
+    SessionDetailSerializer,
+    SessionListItemSerializer,
+    SessionTitleUpdateSerializer,
+)
+from chat_sessions.services import (
+    delete_session,
+    get_session_for_user,
+    list_sessions_for_user,
+    update_session_title,
+)
 from file_processing.services.upload_service import (
     FILE_TOO_LARGE_ERROR,
     MAX_FILE_SIZE,
@@ -705,6 +716,71 @@ def download_csv(request, file_id):
         filename=download_name,
         content_type=artifact["content_type"],
     )
+
+
+@api_view(["GET"])
+@permission_classes([IsAuthenticated, IsVerifiedUser])
+def session_list(request):
+    sessions = list(list_sessions_for_user(request.user))
+    serializer = SessionListItemSerializer(sessions, many=True)
+    return Response(
+        {
+            "count": len(sessions),
+            "results": serializer.data,
+        },
+        status=status.HTTP_200_OK,
+    )
+
+
+@api_view(["GET"])
+@permission_classes([IsAuthenticated, IsVerifiedUser])
+def session_detail(request, session_id):
+    session = get_session_for_user(request.user, session_id)
+    if session is None:
+        return Response({"detail": "Not found."}, status=status.HTTP_404_NOT_FOUND)
+
+    return Response(SessionDetailSerializer(session).data, status=status.HTTP_200_OK)
+
+
+@api_view(["PATCH"])
+@permission_classes([IsAuthenticated, IsVerifiedUser])
+def session_update(request, session_id):
+    session = get_session_for_user(request.user, session_id)
+    if session is None:
+        return Response({"detail": "Not found."}, status=status.HTTP_404_NOT_FOUND)
+
+    serializer = SessionTitleUpdateSerializer(data=request.data)
+    if not serializer.is_valid():
+        return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
+
+    updated_session = update_session_title(session, serializer.validated_data["title"])
+    return Response(
+        SessionListItemSerializer(updated_session).data,
+        status=status.HTTP_200_OK,
+    )
+
+
+@api_view(["DELETE"])
+@permission_classes([IsAuthenticated, IsVerifiedUser])
+def session_delete(request, session_id):
+    session = get_session_for_user(request.user, session_id)
+    if session is None:
+        return Response({"detail": "Not found."}, status=status.HTTP_404_NOT_FOUND)
+
+    delete_session(session)
+    return Response(status=status.HTTP_204_NO_CONTENT)
+
+
+@api_view(["GET", "PATCH", "DELETE"])
+@permission_classes([IsAuthenticated, IsVerifiedUser])
+def session_resource(request, session_id):
+    if request.method == "GET":
+        return session_detail(request, session_id)
+    if request.method == "PATCH":
+        return session_update(request, session_id)
+    return session_delete(request, session_id)
+
+
 
 
 @require_GET
